test(app): keep lifecycle tests in resource_app_test.go
